[修改]更正体力条tick恢复异常问题

+++ Source/PeachOnline/Core/BaseCharacter.h
@@ -70,7 +70,7 @@ public:
 
 	
 	UPROPERTY(BlueprintReadOnly)
-	int RunEnergy;
+	float RunEnergy;
 	
 	UPROPERTY(BlueprintReadOnly)
 	WalkState PlayerWalkState;
